l4re-base-24.08.0

+++ src/l4/pkg/l4re-core/uclibc/lib/contrib/uclibc/libm/nan.c
@@ -1,4 +1,3 @@
-/* vi: set sw=4 ts=4: */
 /*
  * Copyright (C) 2002 by Erik Andersen <andersen@uclibc.org>
  *
@@ -29,7 +28,7 @@ double nan (const char *tagp)
 		sprintf (buf, "NAN(%s)", tagp);
 		return strtod (buf, NULL);
 	}
-	return NAN;
+	return (double)NAN;
 }
 libm_hidden_def(nan)
 
@@ -54,7 +53,7 @@ long double nanl (const char *tagp)
 		sprintf (buf, "NAN(%s)", tagp);
 		return strtold (buf, NULL);
 	}
-	return NAN;
+	return (long double)NAN;
 }
 libm_hidden_def(nanl)
 #endif
